rename `atoms` -> `location`
