+++ src/bloqade/location/__init__.py
@@ -1,6 +1,6 @@
 from .base import AtomArrangement
 from .bravais import Chain, Square, Rectangular, Honeycomb, Triangular, Lieb, Kagome
-from .list import ListOfPositions
+from .list import ListOfLocations
 
 __all__ = [
     "AtomArrangement",
@@ -12,5 +12,5 @@ __all__ = [
     "Lieb",
     "Kagome",
     "BoundedBravais",
-    "ListOfPositions",
+    "ListOfLocations",
 ]
