+++ tests/_test_issue.py
@@ -1,8 +1,8 @@
 from bloqade.ir import Linear
-import bloqade.atoms as atoms
+import bloqade.location as location
 
 (
-    atoms.Square(3)
+    location.Square(3)
     .rydberg.detuning.uniform.apply(Linear(start=1.0, stop="x", duration=3.0))
     .location(2)
     .scale(3.0)
@@ -10,7 +10,7 @@ import bloqade.atoms as atoms
 )
 
 (
-    atoms.Square(3)
+    location.Square(3)
     .rydberg.detuning.uniform.apply(Linear(start=1.0, stop="x", duration=3.0))
     .location(2)
     .scale(3.0)
